refactor: buildApp
